refactor(workspace): move health + rules to claudectl-core (closes #276)
Third step of the workspace-refactor epic (#279). Both modules were
proposed for core ownership in the epic body — they're pure logic over
session state and brain decision inputs, called from TUI display, the
brain's decision input, and (now) the runtime trait wrappers.

Moves:
* src/health.rs → crates/claudectl-core/src/health.rs
* src/rules.rs  → crates/claudectl-core/src/rules.rs

HealthThresholds lived in src/config.rs but it's a struct the health
module owns. Moved with health into core; config.rs `pub use`s it back
so binary callers (and the TOML parser in RawHealthThresholds) keep
working unchanged.

Re-export pattern matches #273: `src/lib.rs` and `src/main.rs` now
include `health, rules` in their `pub use claudectl_core::{...}`
list so every existing `crate::health::*` / `crate::rules::*` path in
the binary keeps resolving without rewriting 30+ call sites.

Build verification:
- cargo check (default + bus + coord) — clean
- cargo build -p claudectl-core (standalone) — clean (100 tests pass,
  up from 61: health + rules added ~39 tests to core)
- cargo clippy --all-targets [--features bus] -- -D warnings — clean
- cargo test --features bus passes (1292 tests; the count moves
  between crates but no tests are lost)

No behavior change. Pure mechanical move.
