+++ crates/claudectl-core/src/rules.rs
